% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mie_cross_section.R
\name{cross_section}
\alias{cross_section}
\title{Scattering and extinction cross-sections}
\usage{
cross_section(radius, lambda, np, nm)
}
\arguments{
\item{radius}{A numeric of the particle radius.}

\item{lambda}{A numeric of the light's wavelength.}

\item{np}{A numeric or complex of the refractive indix of the particle.}

\item{nm}{A numeric or complex of the refractive indix of the medium.}
}
\description{
Calculate scattering and extinction cross-section.
* Csca: scattering cross-section
* Cext: extinction cross-section
}
\examples{
cross_section(50*10^-9, 0.55*10^-6, 0.57+2.45i, 1.33)
cross_section(c(50, 100, 500)*10^-9, 0.55*10^-6, 0.57+2.45i, 1.33)

}
