% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mie_cross_section.R
\name{cross_section_sca}
\alias{cross_section_sca}
\title{Scattering cross-section}
\usage{
cross_section_sca(radius, lambda, np, nm)
}
\arguments{
\item{radius}{A numeric of the particle radius.}

\item{lambda}{A numeric of the light's wavelength.}

\item{np}{A numeric or complex of the refractive indix of the particle.}

\item{nm}{A numeric or complex of the refractive indix of the medium.}
}
\description{
Calculate the scattering cross-section.
}
\examples{
cross_section_sca(50*10^-9, 0.55*10^-6, 0.57+2.45i, 1.33)
cross_section_sca(c(50, 100, 500)*10^-9, 0.55*10^-6, 0.57+2.45i, 1.33)

}
