% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extinction_efficiency.R
\name{ext_efficiency}
\alias{ext_efficiency}
\title{Extinction efficiency}
\usage{
ext_efficiency(radius, lambda, np, nm)
}
\arguments{
\item{radius}{A numeric of the particle radius.}

\item{lambda}{A numeric of the light's wavelength.}

\item{np}{A numeric or complex of the refractive indix of the particle.}

\item{nm}{A numeric or complex of the refractive indix of the medium.}
}
\description{
Calculate the extinction efficiency.
}
\examples{
ext_efficiency(50*10^-9, 0.55*10^-6, 0.57+2.45i, 1.33)
ext_efficiency(c(50, 100, 500)*10^-9, 0.55*10^-6, 0.57+2.45i, 1.33)

}
