% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mie_scattering.R
\name{miescatteR}
\alias{miescatteR}
\title{Calculate the intensity of scattered light}
\usage{
miescatteR(radius, lambda, np, nm, nang = 360)
}
\arguments{
\item{radius}{A numeric of the particle radius.}

\item{lambda}{A numeric of the light's wavelength.}

\item{np}{A numeric or complex of the refractive indix of the particle.}

\item{nm}{A numeric or complex of the refractive indix of the medium.}

\item{nang}{A integer of the angle division number.}
}
\description{
Calculate the intensity of scattered light
}
\details{
\deqn{I(\theta,\varphi) = cos^2\varphi I_\parallel (\theta) + sin^2\varphi I_\perp(\theta)}

\deqn{
I_\parallel (\theta)
 = (\frac{\lambda}{2 \pi r})^2|\sum^\infty_{n=1} \frac{2n+1}{n(n+1)}(a_n \pi_n(cos\theta) + b_n \tau_n(cos\theta))|^2
}
\deqn{
I_\perp (\theta)
 = (\frac{\lambda}{2 \pi r})^2|\sum^\infty_{n=1} \frac{2n+1}{n(n+1)}(a_n \tau_n(cos\theta) + b_n \pi_n(cos\theta))|^2
}
}
\examples{
miescatteR(0.26*10^-6, 0.55*10^-6, 1.33+(1i*10^-8), 1.0, 360)
miescatteR( 50*10^-9, 0.55*10^-6, 0.57+2.45i, 1.33, 360)
miescatteR(100*10^-9, 0.55*10^-6, 0.57+2.45i, 1.33, 360)
miescatteR(500*10^-9, 0.55*10^-6, 0.57+2.45i, 1.33, 360)

}
