% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Riccati_Bessel_funcion.R
\name{psinz}
\alias{psinz}
\alias{xinz}
\alias{psindz}
\alias{xindz}
\title{Riccati-Bessel Funcions}
\usage{
psinz(z, nu, nSeq = 1)

xinz(z, nu, nSeq = 1)

psindz(z, nu, nSeq = 1)

xindz(z, nu, nSeq = 1)
}
\arguments{
\item{z}{A complex or numeric vector.}

\item{nu}{A numeric.}

\item{nSeq}{A positive integer. if >1, the result has \eqn{j_{nu}}, \eqn{j_{nu+1}}, ..., \eqn{j_{nu+nSeq-1}}}
}
\description{
The special functions mostly used in calculating mie scattering.
}
\details{
\eqn{\psi_n(z)} and \eqn{\xi_n(z)}, called Riccati-Bessel functions, are the special functions which slightly differ from spherical Bessel functions.

\deqn{\psi_n(z) = zj_n(z)}
\deqn{\xi_n(z) = zh^{(1)}_n(z)}

These functions satisfy the differential equation;
\deqn{x^2\frac{d^2y}{dx^2} + (x^2 - n(n+1))y = 0}

\eqn{\psi'_n(z)} and \eqn{\xi'_n(z)} are written like these by using recurrence formula;

\deqn{\psi'_n(z) = j_{n-1}(z) - \frac{n}{z}j_n(z)}
\deqn{\xi'_n(z) = h^{(1)}_{n-1}(z) - \frac{n}{z}h^{(1)}_n(z)}
}
\examples{
psinz(1, 1, 10)
xinz(1, 1, 10)
psindz(1, 1, 10)
xindz(1, 1, 10)

}
