% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_Bessels.R
\name{spherical_Bessel}
\alias{spherical_Bessel}
\alias{spherical_BesselJ}
\alias{spherical_BesselY}
\alias{spherical_BesselI}
\alias{spherical_BesselK}
\alias{spherical_BesselH}
\title{Spherical Bessel and Hankel functions.}
\usage{
spherical_BesselJ(z, nu, expon.scaled = FALSE, nSeq = 1, verbose = 0)

spherical_BesselY(z, nu, expon.scaled = FALSE, nSeq = 1, verbose = 0)

spherical_BesselI(z, nu, expon.scaled = FALSE, nSeq = 1, verbose = 0)

spherical_BesselK(z, nu, expon.scaled = FALSE, nSeq = 1, verbose = 0)

spherical_BesselH(m, z, nu, expon.scaled = FALSE, nSeq = 1, verbose = 0)
}
\arguments{
\item{z}{A complex or numeric vector.}

\item{nu}{A numeric.}

\item{expon.scaled}{Logical. The result should be scaled by an exponential factor,
typically to avoid under- or over-flow.}

\item{nSeq}{A positive integer. if >1, the result has \eqn{j_{nu}}, \eqn{j_{nu+1}}, ..., \eqn{j_{nu+nSeq-1}}}

\item{verbose}{A integer (default is 0). Indicating the level of verbosity notably from C code.}

\item{m}{1 or 2. Indicate the kind of Hankel function.}
}
\description{
Extend Bessel functions, modified Bessel functions and Hankel function to spherical functions.
}
\details{
The differential equation

\deqn{x^2 \frac{d^2y}{dx^2} + 2x \frac{dy}{dx} + (x^2 - n(n+1))y = 0}

has two linearly independent solutions which are called the spherical Bessel functions.

\deqn{j_n(x) = \sqrt{\frac{\pi}{2x}} J_{n+\frac{1}{2}}(x)}
\deqn{y_n(x) = \sqrt{\frac{\pi}{2x}} Y_{n+\frac{1}{2}}(x)}

These are the spherical Hankel function.

\deqn{h^{(1)}_n(x) = j_n(x) + iy_n(x)}
\deqn{h^{(2)}_n(x) = j_n(x) - iy_n(x)}
}
\examples{
spherical_BesselJ(3+5i, 1)
spherical_BesselJ(3+5i, 1, nSeq = 4)
spherical_BesselJ(c(3+5i, 1+2i, 7+6i), 1)
spherical_BesselJ(c(3+5i, 1+2i, 7+6i), 1, nSeq = 4)

}
